fix: move ee tasks to celery.py
